Updated calculation for taxable income (SELIC).

+++ src/main/java/br/com/bvinvestimentos/calculadora/api/selic/SelicAdapter.java
@@ -1,5 +1,7 @@
 package br.com.bvinvestimentos.calculadora.api.selic;
 
+import java.math.BigDecimal;
+
 public interface SelicAdapter {
-	public Double getSelic();
+	public BigDecimal getSelic();
 }
